+++ khanadat_proto/exec/exec_pipe_utils.c
@@ -6,7 +6,7 @@
 /*   By: khanadat <khanadat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 21:44:49 by khanadat          #+#    #+#             */
-/*   Updated: 2025/10/14 23:23:21 by khanadat         ###   ########.fr       */
+/*   Updated: 2025/10/15 13:34:35 by khanadat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,7 @@
 #include "libft.h"
 
 static t_node	*connect_cmd(t_node *node);
+static void		exec_pipe_child(t_mini *mini, t_node *before, t_node *node);
 
 t_node	*set_cmd_order(t_node *node)
 {
@@ -31,23 +32,22 @@ t_node	*set_cmd_order(t_node *node)
 
 void	start_pipe_group(t_mini *mini, t_node *node)
 {
-	bool	before;
+	t_node	*before;
 
 	before = NULL;
 	while (node)
 	{
-		if (!before)
+		if (node->rhs)
 			pipe(node->cmd->pfd);
 		node->cmd->pid = fork();
 		if (node->cmd->pid < 0)
 			systemcall_minishell_exit(mini, "fork");
 		if (node->cmd->pid == 0)
+			exec_pipe_child(mini, before, node);
+		if (0 < node->cmd->pid && before)
 		{
-			if (before)
-			{
-				close(node->cmd->pfd[1]);
-			}
-			exec_cmd(mini, node);
+			close(before->cmd->pfd[0]);
+			close(before->cmd->pfd[1]);
 		}
 		before = node;
 		node = node->rhs;
@@ -60,7 +60,11 @@ void	wait_pipe_group(t_mini *mini, t_node *node)
 
 	while (node)
 	{
-		waitpid(node->cmd->pid, &status, 0);
+		if (waitpid(node->cmd->pid, &status, 0) < 0)
+			systemcall_minishell_exit(mini, "waitpid");
+		if (!node->rhs)
+			mini->is_pipe = false;
+		catch_signal(status, mini);
 		node = node->rhs;
 	}
 }
@@ -81,4 +85,25 @@ static t_node	*connect_cmd(t_node *node)
 		return (node->rhs);
 	}
 	return (NULL);
+}
+
+static void	exec_pipe_child(t_mini *mini, t_node *before, t_node *node)
+{
+	if (before)
+	{
+		close(before->cmd->pfd[1]);
+		if (dup2(before->cmd->pfd[0], STDIN_FILENO) < 0)
+			systemcall_minishell_exit(mini, "dup2");
+		close(before->cmd->pfd[0]);
+	}
+	if (node->rhs)
+	{
+		close(node->cmd->pfd[0]);
+		if (dup2(node->cmd->pfd[1], STDOUT_FILENO) < 0)
+			systemcall_minishell_exit(mini, "dup2");
+		close(node->cmd->pfd[1]);
+	}
+	if (mini->signaled)
+		normal_minishell_exit(mini, NULL, NULL, ft_atoi(mini->status));
+	exec_cmd(mini, node);
 }
