+++ khanadat_proto/exec/exec_utils0.c
@@ -6,7 +6,7 @@
 /*   By: khanadat <khanadat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/09 09:45:04 by khanadat          #+#    #+#             */
-/*   Updated: 2025/10/14 21:16:39 by khanadat         ###   ########.fr       */
+/*   Updated: 2025/10/15 13:21:07 by khanadat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
