+++ khanadat_proto/utils/minishell_utils.h
@@ -6,7 +6,7 @@
 /*   By: khanadat <khanadat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/09 12:45:32 by khanadat          #+#    #+#             */
-/*   Updated: 2025/10/14 21:15:50 by khanadat         ###   ########.fr       */
+/*   Updated: 2025/10/15 13:20:56 by khanadat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
